pyproject.toml
[project]
name = "GameGrid"
description = "A flask wrapper for your WebGL games"

[tool.poetry]
name = "GameGrid"
version = "0.1.0"
description = "A flask wrapper for your WebGL games"
authors = ["PixelWelt"]
license = "MIT"
readme = "README.md"
keywords = ["flask", "webgl", "arcade", "games"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
blinker= "1.9.0"
click= "8.1.8"
colorama= "0.4.6"
Flask= "3.1.0"
Flask-Cors= "5.0.0"
itsdangerous= "2.2.0"
Jinja2= "3.1.5"
loguru= "0.7.3"
MarkupSafe= "3.0.2"
poetry = ">=2.1.0"
gunicorn = "23.0.0"
poetry-core = ">=2.1.0"
PyYAML= "6.0.2"
Werkzeug= "3.1.3"
win32_setctime= "1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project.urls]
Repository = "https://github.com/PixelWelt/GameGrid"
